Flip the endianness of the function flags (#1616)
This value was mostly pass-through: we read it
and wrote it back, without interpretting it, so
it was easy for this bug to sneak through.

The one catch we didn't handle is the debug
flag is always true in practice, but with
ES2015 we now see the incorrect flag triggering
incorrect behavior.

This was annoying to isolate.

Co-authored-by: Jesse Wilson <jwilson@squareup.com>